+++ src/Libraries/SmartStore.Data/Migrations/202002241354543_ProductCondition.cs
@@ -1,0 +1,20 @@
+namespace SmartStore.Data.Migrations
+{
+    using System;
+    using System.Data.Entity.Migrations;
+    
+    public partial class ProductCondition : DbMigration
+    {
+        public override void Up()
+        {
+            AddColumn("dbo.Product", "Condition", c => c.Int(nullable: false));
+            CreateIndex("dbo.ProductAttribute", "DisplayOrder");
+        }
+        
+        public override void Down()
+        {
+            DropIndex("dbo.ProductAttribute", new[] { "DisplayOrder" });
+            DropColumn("dbo.Product", "Condition");
+        }
+    }
+}